Fix insecure source URI error
Why:

* CodeClimate marks the error `Insecure Source URI found:
git://github.com/middleman-contrib/middleman-deploy.git` in
`Gemfile.lock`. According to bundler/bundler#3324 this happens because
the default for GitHub linked gems is to include the link using the
Git scheme.
* Since the `middleman-deploy` gem is not up to date, we need to use the
current version on master. Relying on someone else's repository directly
is brittle. We forked the repository for safety so we need to change it
to our branch.

This change addresses the problem by:

* Changing the `middleman-deploy` gem to be included from our repository
using HTTPS.